Coalesce multiple updates to the service map while waiting to be scheduled by the main looper.
There could be a flurry of onServiceLost() or results from resolve()
in the background threads. Currently the code posts a runnable to the
main thread for every update, creating a similar flurry of calls to
onServicesChanged().

Instead post a single runnable which when scheduled by the main looper
has the latest copy.